requirements.txt
# tkinter
# pytesseract
# Pillow
# keyboard
# difflib

pytesseract==0.3.10
Pillow==9.5.0
keyboard==0.13.5
pyinstaller==6.12.0